+++ examples/main/shared/editor/EditorPlugins.jsx
@@ -1,13 +1,25 @@
-import React from 'react';
-import { createLinkPlugin, LINK_TYPE } from 'wix-rich-content-plugin-link';
-import { createLineSpacingPlugin, LINE_SPACING_TYPE } from 'wix-rich-content-plugin-line-spacing';
-import { createHashtagPlugin, HASHTAG_TYPE } from 'wix-rich-content-plugin-hashtag';
-// import { createExternalEmojiPlugin, EXTERNAL_EMOJI_TYPE } from 'wix-rich-content-plugin-emoji';
-import { createImagePlugin, IMAGE_TYPE } from 'wix-rich-content-plugin-image';
-import { createGalleryPlugin, GALLERY_TYPE } from 'wix-rich-content-plugin-gallery';
-import { createVideoPlugin, VIDEO_TYPE } from 'wix-rich-content-plugin-video';
-import { createHtmlPlugin, HTML_TYPE } from 'wix-rich-content-plugin-html';
-import { createDividerPlugin, DIVIDER_TYPE } from 'wix-rich-content-plugin-divider';
+import React from "react";
+import { createLinkPlugin, LINK_TYPE } from "wix-rich-content-plugin-link";
+import {
+  createLineSpacingPlugin,
+  LINE_SPACING_TYPE
+} from "wix-rich-content-plugin-line-spacing";
+import {
+  createHashtagPlugin,
+  HASHTAG_TYPE
+} from "wix-rich-content-plugin-hashtag";
+import { createEmojiPlugin } from "wix-rich-content-plugin-emoji";
+import { createImagePlugin, IMAGE_TYPE } from "wix-rich-content-plugin-image";
+import {
+  createGalleryPlugin,
+  GALLERY_TYPE
+} from "wix-rich-content-plugin-gallery";
+import { createVideoPlugin, VIDEO_TYPE } from "wix-rich-content-plugin-video";
+import { createHtmlPlugin, HTML_TYPE } from "wix-rich-content-plugin-html";
+import {
+  createDividerPlugin,
+  DIVIDER_TYPE
+} from "wix-rich-content-plugin-divider";
 import {
   createExternalMentionsPlugin,
   EXTERNAL_MENTIONS_TYPE,
@@ -67,7 +79,6 @@ export const editorPlugins = [
   createVideoPlugin,
   createHtmlPlugin,
   createDividerPlugin,
-  // createExternalEmojiPlugin,
   createLineSpacingPlugin,
   createLinkPlugin,
   createHashtagPlugin,
@@ -80,19 +91,20 @@ export const editorPlugins = [
   createFileUploadPlugin,
   createButtonPlugin,
   createTextColorPlugin,
+  createEmojiPlugin,
   createTextHighlightPlugin,
-  createBlockDndPlugin,
+  createBlockDndPlugin
 ];
 
 const themeColors = {
-  color1: '#ffffff',
-  color2: '#303030',
-  color3: '#3a54b4',
-  color4: '#bfad80',
-  color5: '#bf695c',
-  color6: '#f7f7f7',
-  color7: '#000000',
-  color8: '#9a87ce',
+  color1: "#ffffff",
+  color2: "#303030",
+  color3: "#3a54b4",
+  color4: "#bfad80",
+  color5: "#bf695c",
+  color6: "#f7f7f7",
+  color7: "#000000",
+  color8: "#9a87ce"
 };
 
 const buttonDefaultPalette = ['#FEFDFD', '#D5D4D4', '#ABCAFF', '#81B0FF', '#0261FF', '#0141AA'];
@@ -102,11 +114,11 @@ let userButtonBorderColors = [...buttonDefaultPalette];
 
 const getLinkPanelDropDownConfig = () => {
   const getItems = () => {
-    casual.define('item', function() {
+    casual.define("item", function() {
       return {
         value: casual.url,
         label: casual.catch_phrase,
-        date: casual.date('DD/MM/YY'),
+        date: casual.date("DD/MM/YY")
       };
     });
 
@@ -122,7 +134,9 @@ const getLinkPanelDropDownConfig = () => {
     <Highlighter
       searchWords={[searchWords]}
       textToHighlight={textToHighlight}
-      highlightTag={({ children }) => <strong className="highlighted-text">{children}</strong>}
+      highlightTag={({ children }) => (
+        <strong className="highlighted-text">{children}</strong>
+      )}
     />
   );
 
@@ -134,20 +148,26 @@ const getLinkPanelDropDownConfig = () => {
     itemHeight: 40,
     itemToString: item => item.value,
     formatMenuItem: (item, input) => (
-      <div style={{ display: 'flex', justifyContent: 'space-between', padding: '10px' }}>
+      <div
+        style={{
+          display: "flex",
+          justifyContent: "space-between",
+          padding: "10px"
+        }}
+      >
         <span
           style={{
-            whiteSpace: 'nowrap',
-            overflow: 'hidden',
-            textOverflow: 'ellipsis',
-            paddingRight: '10px',
+            whiteSpace: "nowrap",
+            overflow: "hidden",
+            textOverflow: "ellipsis",
+            paddingRight: "10px"
           }}
         >
           {wordHighlighter(item.label, input)}
         </span>
         <span>{item.date}</span>
       </div>
-    ),
+    )
   };
 };
 
@@ -190,11 +210,11 @@ export const config = {
     //   },
     // },
     imageEditorWixSettings: {
-      initiator: 'some-initiator',
+      initiator: "some-initiator",
       siteToken:
-        'JWS.eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCIsImtpZCI6Im5FUXljQzlOIn0.eyJpYXQiOjE1Njc1MjY3NzQsImRhdGEiOiJ7XCJ1c2VySWRcIjpcIjE5YTY0YTRjLWVlZTAtNGYxNC1iNjI3LTY3MmQ1ZjE2OGJkNFwiLFwibWV0YXNpdGVJZFwiOlwiNTM4ZmE2YzYtYzk1My00Y2RkLTg2YzQtNGI4NjlhZWNmOTgwXCJ9IiwiZXhwIjoxNTY4NzM2Mzc0fQ.n21OxIzSbqi8N3v30b6cIxMdshBnkkf2WQLWEFVXsLk',
-      metaSiteId: '538fa6c6-c953-4cdd-86c4-4b869aecf980',
-      mediaRoot: 'some-mediaRoot',
+        "JWS.eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCIsImtpZCI6Im5FUXljQzlOIn0.eyJpYXQiOjE1Njc1MjY3NzQsImRhdGEiOiJ7XCJ1c2VySWRcIjpcIjE5YTY0YTRjLWVlZTAtNGYxNC1iNjI3LTY3MmQ1ZjE2OGJkNFwiLFwibWV0YXNpdGVJZFwiOlwiNTM4ZmE2YzYtYzk1My00Y2RkLTg2YzQtNGI4NjlhZWNmOTgwXCJ9IiwiZXhwIjoxNTY4NzM2Mzc0fQ.n21OxIzSbqi8N3v30b6cIxMdshBnkkf2WQLWEFVXsLk",
+      metaSiteId: "538fa6c6-c953-4cdd-86c4-4b869aecf980",
+      mediaRoot: "some-mediaRoot"
     },
     onImageEditorOpen: () => console.log('Media Studio Launched'),
     // toolbar: {
@@ -216,11 +236,12 @@ export const config = {
     // },
   },
   [HASHTAG_TYPE]: {
-    createHref: decoratedText => `/search/posts?query=${encodeURIComponent('#')}${decoratedText}`,
+    createHref: decoratedText =>
+      `/search/posts?query=${encodeURIComponent("#")}${decoratedText}`,
     onClick: (event, text) => {
       event.preventDefault();
       console.log(`'${text}' hashtag clicked!`);
-    },
+    }
   },
   [HTML_TYPE]: {
     htmlIframeSrc: `${getBaseUrl()}/static/html-plugin-embed.html`,
@@ -239,33 +260,33 @@ export const config = {
     repositionSuggestions: true,
     visibleItemsBeforeOverflow: 5,
     popoverComponent: <div />,
-    handleDropdownOpen: () => console.log('mentions dropdown opened'),
+    handleDropdownOpen: () => console.log("mentions dropdown opened"),
     onMentionClick: mention => console.log({ mention }),
-    handleDropdownClose: () => console.log('mentions dropdown closed'),
+    handleDropdownClose: () => console.log("mentions dropdown closed"),
     getMentions: searchQuery =>
       new Promise(resolve =>
         setTimeout(
           () =>
             resolve([
               { name: searchQuery, slug: searchQuery },
-              { name: 'Test One', slug: 'testone' },
-              { name: 'Test One.1', slug: 'testone1' },
-              { name: 'Test One.2', slug: 'testone2' },
-              { name: 'Test One.3', slug: 'testone3' },
-              { name: 'Test One.4', slug: 'testone4' },
-              { name: 'Test One.5', slug: 'testone5' },
-              { name: 'Test One.6', slug: 'testone6' },
-              { name: 'Test One.7', slug: 'testone7' },
-              { name: 'Test One.8', slug: 'testone8' },
+              { name: "Test One", slug: "testone" },
+              { name: "Test One.1", slug: "testone1" },
+              { name: "Test One.2", slug: "testone2" },
+              { name: "Test One.3", slug: "testone3" },
+              { name: "Test One.4", slug: "testone4" },
+              { name: "Test One.5", slug: "testone5" },
+              { name: "Test One.6", slug: "testone6" },
+              { name: "Test One.7", slug: "testone7" },
+              { name: "Test One.8", slug: "testone8" },
               {
-                name: 'Test Two',
-                slug: 'testwo',
-                avatar: 'https://via.placeholder.com/100x100?text=Image=50',
-              },
+                name: "Test Two",
+                slug: "testwo",
+                avatar: "https://via.placeholder.com/100x100?text=Image=50"
+              }
             ]),
           250
         )
-      ),
+      )
   },
   [LINE_SPACING_TYPE]: {
     // toolbar: {
@@ -274,11 +295,11 @@ export const config = {
     //   },
     // },
     defaultSpacing: {
-      'line-height': '1.5',
-      'padding-top': '2px',
-      'padding-bottom': '3px',
+      "line-height": "1.5",
+      "padding-top": "2px",
+      "padding-bottom": "3px"
     },
-    onUpdate: spacing => console.log(LINE_SPACING_TYPE, spacing),
+    onUpdate: spacing => console.log(LINE_SPACING_TYPE, spacing)
   },
   [LINK_TYPE]: {
     // toolbar: {
@@ -301,17 +322,18 @@ export const config = {
     },
     //Here you can call your custom video upload functionality (comment function to disable custom upload)
     handleFileSelection: (updateEntity, removeEntity) => {
-      console.log('consumer wants to upload custom video');
+      console.log("consumer wants to upload custom video");
       const videoWithAbsoluteUrl = {
-        url: 'http://mirrors.standaloneinstaller.com/video-sample/jellyfish-25-mbps-hd-hevc.mp4',
+        url:
+          "http://mirrors.standaloneinstaller.com/video-sample/jellyfish-25-mbps-hd-hevc.mp4"
       };
       const videoWithRelativeUrl = {
-        pathname: 'video/441c23_84f5c058e5e4479ab9e626cd5560a21b/file',
+        pathname: "video/441c23_84f5c058e5e4479ab9e626cd5560a21b/file",
         thumbnail: {
-          pathname: 'media/441c23_84f5c058e5e4479ab9e626cd5560a21bf000.jpg',
+          pathname: "media/441c23_84f5c058e5e4479ab9e626cd5560a21bf000.jpg",
           height: 1080,
-          width: 1920,
-        },
+          width: 1920
+        }
       };
       // You can provide either absolute or relative URL.
       // If relative URL is provided, a function 'getVideoUrl' will be invoked to form a full URL.
@@ -319,7 +341,7 @@ export const config = {
       setTimeout(() => {
         updateEntity({ data: videoToUpload });
         //updateEntity({ error: { msg: 'Upload Failed' } });
-        console.log('consumer uploaded ', videoToUpload);
+        console.log("consumer uploaded ", videoToUpload);
       }, 500);
     },
     // handleFileUpload: (file, updateEntity, removeEntity) => {
@@ -347,7 +369,7 @@ export const config = {
     enableCustomUploadOnMobile: true,
     // Function is invoked when rendering video which has relative URL.
     // You should take the pathname and form a full URL.
-    getVideoUrl: src => `https://video.wixstatic.com/${src.pathname}`,
+    getVideoUrl: src => `https://video.wixstatic.com/${src.pathname}`
   },
   [GIPHY_TYPE]: {
     giphySdkApiKey: process.env.GIPHY_API_KEY,
@@ -365,12 +387,13 @@ export const config = {
     minHeight: 100,
     maxHeight: 1000,
     mapSettings: {
-      address: 'Wix HQ, Nemal Tel Aviv Street, Tel Aviv-Yafo, Israel',
-      locationDisplayName: 'Wix HQ, Nemal Tel Aviv Street, Tel Aviv-Yafo, Israel',
+      address: "Wix HQ, Nemal Tel Aviv Street, Tel Aviv-Yafo, Israel",
+      locationDisplayName:
+        "Wix HQ, Nemal Tel Aviv Street, Tel Aviv-Yafo, Israel",
       lat: 32.097235,
       lng: 34.77427,
       zoom: 18,
-      mode: 'roadmap',
+      mode: "roadmap",
       isMarkerShown: true,
       isZoomControlShown: true,
       isStreetViewControlShown: true,
@@ -391,16 +414,16 @@ export const config = {
     accept: '*',
     onFileSelected: (file, updateEntity) => {
       const name = file.name;
-      const filenameParts = name.split('.');
+      const filenameParts = name.split(".");
       const type = filenameParts[filenameParts.length - 1];
 
       const data = {
         name,
         type,
-        url: '',
+        url: ""
       };
       setTimeout(() => updateEntity({ data }), 1000);
-    },
+    }
     // handleFileSelection: updateEntity => {
     //   const filenames = ['image.jpg', 'document.pdf', 'music.mp3'];
     //   const name = filenames[Math.floor(Math.random() * filenames.length)];
@@ -461,7 +484,7 @@ export const config = {
     styleSelectionPredicate,
     customStyleFn: customForegroundStyleFn,
     onColorAdded: color => (userColors = [color, ...userColors]),
-    getUserColors: () => userColors,
+    getUserColors: () => userColors
   },
   uiSettings,
   getToolbarSettings: ({ pluginButtons, textButtons }) => [
@@ -598,5 +621,5 @@ export const config = {
     //     desktop: () => InlineToolbarDecoration
     //   })
     // }
-  ],
+  ]
 };