+++ packages/plugin-emoji/web/src/index.js
@@ -1,2 +1,2 @@
-export { createExternalEmojiPlugin } from './createEmojiPlugin';
-export { EXTERNAL_EMOJI_TYPE } from './types';
+export { createEmojiPlugin } from './createEmojiPlugin';
+export { EMOJI_TYPE } from './constants';