+++ packages/plugin-emoji/web/src/types.js
@@ -1,1 +1,0 @@
-export const EXTERNAL_EMOJI_TYPE = 'wix-draft-plugin-external-emoji';